+++ docs/web3.utils.rst
@@ -5,6 +5,21 @@ Utils
 
 The ``utils`` module houses public utility and helper functions.
 
+ABI
+---
+
+.. py:method:: Utils.get_abi_input_names(abi)
+
+    Return the ``input`` names for an ABI function or event.
+
+
+.. py:method:: Utils.get_abi_output_names(abi)
+
+    Return the ``output`` names an ABI function or event.
+
+
+Exception Handling
+------------------
 
 .. py:method:: Utils.handle_offchain_lookup(offchain_lookup_payload, transaction)
