+++ web3/utils/__init__.py
@@ -1,3 +1,11 @@
+"""
+NOTE: This is a public utility module. Any changes to these utility methods would
+classify as breaking changes.
+"""
+from .abi import (  # NOQA
+    get_abi_input_names,
+    get_abi_output_names,
+)
 from .async_exception_handling import (  # NOQA
     async_handle_offchain_lookup,
 )
